server/python_scripts/requirements.txt
psycopg==3.2.13
psycopg-binary==3.2.13
psycopg-pool==3.2.8
typing-extensions==4.15.0

python-dotenv==1.0.0

requests==2.31.0
nba_api==1.4.1
pandas==2.2.2
